fix: restore github issue build path
