+++ AgentBoard/Services/PRDGenerator-GitHub.swift
@@ -1,133 +1,5 @@
 import Foundation
 
-/// Service for generating PRD (Product Requirements Document) markdown from GitHub Issues
-public final class PRDGenerator: Sendable {
-    
-    public init() {}
-    
-    /// Generate PRD markdown from a GitHub Issue
-    public func generatePRD(from issue: GitHubIssue) -> String {
-        var lines: [String] = []
-        
-        // Title
-        lines.append("# #\(issue.number): \(issue.title)")
-        lines.append("")
-        
-        // Metadata
-        lines.append("## Context")
-        lines.append("")
-        lines.append("- **Issue:** #\(issue.number)")
-        if !issue.labels.isEmpty {
-            let labelNames = issue.labels.map { $0.name }.joined(separator: ", ")
-            lines.append("- **Labels:** \(labelNames)")
-        }
-        if let milestone = issue.milestone {
-            lines.append("- **Milestone:** \(milestone.title)")
-        }
-        if !issue.assignees.isEmpty {
-            let assigneeNames = issue.assignees.map { $0.login }.joined(separator: ", ")
-            lines.append("- **Assignees:** \(assigneeNames)")
-        }
-        lines.append("- **Status:** \(issue.state)")
-        lines.append("")
-        
-        // Description
-        if let body = issue.body, !body.isEmpty {
-            lines.append("## Description")
-            lines.append("")
-            lines.append(body)
-            lines.append("")
-        }
-        
-        // Tasks (extract from issue body)
-        let tasks = extractTasks(from: issue.body ?? "")
-        if !tasks.isEmpty {
-            lines.append("## Tasks")
-            lines.append("")
-            for task in tasks {
-                lines.append("- [ ] \(task)")
-            }
-            lines.append("")
-        }
-        
-        // Acceptance criteria
-        lines.append("## Acceptance Criteria")
-        lines.append("")
-        lines.append("- [ ] All tests pass")
-        lines.append("- [ ] Code review completed")
-        lines.append("- [ ] No regressions introduced")
-        lines.append("- [ ] Documentation updated (if needed)")
-        lines.append("")
-        
-        return lines.joined(separator: "\n")
-    }
-    
-    /// Generate combined PRD from multiple GitHub Issues
-    public func generatePRD(from issues: [GitHubIssue]) -> String {
-        var lines: [String] = []
-        
-        lines.append("# PRD: Combined Issues")
-        lines.append("")
-        lines.append("Generated: \(Date())")
-        lines.append("Issues: \(issues.count)")
-        lines.append("")
-        
-        for issue in issues {
-            lines.append("---")
-            lines.append("")
-            lines.append(generatePRD(from: issue))
-            lines.append("")
-        }
-        
-        return lines.joined(separator: "\n")
-    }
-    
-    /// Save PRD to a temporary file and return the path
-    public func savePRD(content: String, issueNumber: Int) -> String {
-        let filename = "PRD-\(issueNumber).md"
-        let tempDir = FileManager.default.temporaryDirectory
-        let prdPath = tempDir.appendingPathComponent(filename)
-        
-        try? content.write(to: prdPath, atomically: true, encoding: .utf8)
-        
-        return prdPath.path
-    }
-    
-    // MARK: - Private Helpers
-    
-    private func extractTasks(from body: String) -> [String] {
-        var tasks: [String] = []
-        
-        // Extract checkbox items: - [ ] Task description
-        let checkboxPattern = "- \\[ \\] (.+)"
-        if let regex = try? NSRegularExpression(pattern: checkboxPattern, options: .caseInsensitive) {
-            let matches = regex.matches(in: body, range: NSRange(body.startIndex..., in: body))
-            for match in matches {
-                if let range = Range(match.range(at: 1), in: body) {
-                    let task = String(body[range]).trimmingCharacters(in: .whitespaces)
-                    if !task.isEmpty {
-                        tasks.append(task)
-                    }
-                }
-            }
-        }
-        
-        // If no checkboxes found, try to extract from description
-        if tasks.isEmpty {
-            // Look for numbered lists or bullet points
-            let lines = body.components(separatedBy: .newlines)
-            for line in lines {
-                let trimmed = line.trimmingCharacters(in: .whitespaces)
-                if trimmed.hasPrefix("- ") || trimmed.hasPrefix("* ") || 
-                   (trimmed.count > 2 && trimmed[trimmed.startIndex].isNumber && trimmed[trimmed.index(after: trimmed.startIndex)] == ".") {
-                    let task = String(trimmed.dropFirst(2)).trimmingCharacters(in: .whitespaces)
-                    if !task.isEmpty {
-                        tasks.append(task)
-                    }
-                }
-            }
-        }
-        
-        return tasks
-    }
-}
+// Canonical PRD generation now lives in AgentBoard/PRDGenerator.swift.
+// This file is intentionally empty to preserve the existing project structure
+// while removing the broken duplicate Bead/GitHub implementation split.
